requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.12.0
mysql-connector-python>=8.3.0
pydantic>=2.6.0
python-dotenv>=1.0.0
